pyproject.toml
[project]
name = "ocelot" # Required
version = "0.4.0" # Required
description = "A toolbox for working with observations of star clusters." # Optional
readme = "README.md" # Optional
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["astronomy", "star cluster", "threading", "development"] # Optional
authors = [
  { name = "Emily Hunt", email = "emily.hunt.physics@gmail.com" }, # Optional
]
maintainers = [
  { name = "Emily Hunt", email = "emily.hunt.physics@gmail.com" }, # Optional
]
classifiers = [ # Optional
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "numpy>1.21.0,<3.0",
  "numba<0.60",  # Todo increase when https://github.com/numba/numba/issues/9753 fix is on main
  "matplotlib>3.4.0,<4.0",
  "scikit-learn>0.24.0,<2.0",
  "scipy>1.0.0,<2.0",
  "pandas>1.0.0,<3.0",
  "pyarrow",
  "astropy>=6.0.0,<7.0",
  "healpy>1.13.0,<2.0",    # Todo consider removing dependency - switch to Astropy's healpy implementation
  "galpy>1.9,<2.0",  # Todo consider removing dependency
  "imf @ git+https://github.com/keflavich/imf",
  "dustmaps",  # Todo is this needed?
  "gaiaunlimited",  # Todo make optional?
  "kepler.py",  # For advanced calculations of the separations of binary stars
]


[project.optional-dependencies]
dev = [
  "mkdocs-material[imaging]",
  "mkdocstrings[python]>=0.18",
  "mkdocs-jupyter",
  "jupyter",  # for working on notebooks
  "check-manifest",
  "pytest",
  "ruff",
  "ezpadova @ git+https://github.com/mfouesneau/ezpadova",  # PARSEC isochrone download
]
docs = [
  "mkdocs-material[imaging]",
  "mkdocstrings[python]>=0.18",
  "mkdocs-jupyter",
] # For building docs on GitHub
test = ["pytest"]

[project.urls] # Optional
"Homepage" = "https://ocelot-docs.org"
"Bug Reports" = "https://github.com/emilyhunt/ocelot/issues"
"Source" = "https://github.com/emilyhunt/ocelot"

# [tool.setuptools]
# package-data = {"sample" = ["*.dat"]}

[tool.pytest.ini_options]
pythonpath = "src"
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
